Update PdfPrintDocument.cs
Fixes problem with printing documents that contain multiple page sizes.  Previously this would render all pages with the page size of the first page of the document.
This attempts to find a closely matching page size in the available page sizes and if so uses the pre-defined matching page size, otherwise it creates a custom page size for the page.